Adding null decision  in ShoppingCartProduct Repository method DELETE

+++ src/main/java/co/edu/eam/disenosoftware/homeautobackend/repositories/ShoppingCartProductRepository.java
@@ -52,7 +52,9 @@ public class ShoppingCartProductRepository {
    */
   public ShoppingCartProduct delete(Long id) {
     ShoppingCartProduct shoppingCartProduct = find(id);
-    em.remove(shoppingCartProduct);
+    if (shoppingCartProduct != null) {
+      em.remove(shoppingCartProduct);
+    }
     return shoppingCartProduct;
   }
